Extract setup wizard window resizing
Move setup wizard resize measurement and scheduling into a dedicated resizer so the window keeps only UI workflow coordination.
