+++ Packages/src/Editor/Presentation/Setup/SetupWizardWindowResizer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3286b30743dc4c4594b96e6ecf4ae6b9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
